Enables closing script tags in Nodes
Previously `Node(:div, setInnerHtml="helloooo!<script>console.log(\"ace\")</script>")` would cause an error during rendering because the `</script>` would close the `<script>` tag that the WebIO.node code lived in, despite the fact that it is in a javascript string (the browser parses it as html, ignoring the surrounding javascript context).

This commit fixes it by escaping the closing script tag before in Julia as "</_script>" and unescaping it back to "</script>" on the js side.